cloud_functions/process_daily_max_wrf/requirements.txt
netCDF4
wrf-python
cartopy
google-cloud-storage
gcsfs
matplotlib
numpy
